chapter_004/src/test/java/ru/job4j/iterator/listtest/SimpleArrayListTest.java
package ru.job4j.iterator.listtest;

import org.junit.Test;
import org.junit.Before;
import ru.job4j.list.SimpleArrayList;

import static org.hamcrest.Matchers.is;
import static org.junit.Assert.assertThat;

public class SimpleArrayListTest {
    private SimpleArrayList<Integer> list;

    @Before
    public void beforeTest() {
        list = new SimpleArrayList<>();
        list.add(1);
        list.add(2);
        list.add(3);
        list.add(4);
        list.add(5);
        list.add(6);
        list.add(7);
        list.add(8);
        list.add(9);
        list.add(10);
    }

    @Test
    public void whenAddThreeElementsThenUseGetOneResultTwo() {
        assertThat(list.get(1), is(2));
    }

    @Test
    public void whenAddThreeElementsThenUseGetSizeResultThree() {
        assertThat(list.getSize(), is(3));
    }

    @Test
    public void whenDeleteThenFirstDeleted() {
        while (list.getSize() > 0)
        list.delete();
        //assertThat(list.get(0), is(2));
    }
}